cassandra_app/requirements.txt
Flask~=2.1.2
pandas~=1.4.2
numpy~=1.22.4
kafka-python==2.0.2